[KTAI-8] ref: Code Separation
