+++ src/shared/interfaces/UserCredentials.interfaces.ts
@@ -1,0 +1,4 @@
+export interface UserCredentials {
+  username: string;
+  password: string;
+}
